refactor: extract shortenPath to shared utility for reuse
