+++ src/renderer/src/components/TabItem.tsx
@@ -8,17 +8,7 @@ import type { NotificationLevel } from '../../../shared/types';
 import { cwdBasename } from '../store/workspace-store';
 import { useCwdStore } from '../store/cwd-store';
 import { useNotificationStore } from '../store/notification-store';
-
-const HOME = window.fleet.homeDir;
-
-function shortenPath(cwd: string): string {
-  const withTilde = cwd.startsWith(HOME) ? '~' + cwd.slice(HOME.length) : cwd;
-  if (withTilde.length <= 30) return withTilde;
-  const parts = withTilde.split('/').filter(Boolean);
-  if (parts.length <= 2) return withTilde;
-  const prefix = withTilde.startsWith('~') ? '~' : '';
-  return `${prefix}/\u2026/${parts.slice(-2).join('/')}`;
-}
+import { shortenPath } from '../lib/shorten-path';
 
 type TabItemProps = {
   id: string;
